+++ apps/carbon/src/types/index.d.ts
@@ -5,3 +5,10 @@
 
 /// <reference types="vite-plugin-svgr/client" />
 /// <reference types="vite/client" />
+
+// TODO: Replace it with the actual type in the following task
+type Extension = {
+  name: string
+  title: string
+  url: string
+}